SF2 spec compliance: fix 11 generator mappings, implement LFO/ModEnv system, MIDI parser robustness
- Fix gen 15/16 (chorus/reverb swap), gen 24 (freqVibLFO), gen 25-30 (ModEnv ADSR),
  gen 31/32 (keynumToModEnv), gen 33/35 (VolEnv delay/hold), gen 39/40 (keynumToVolEnv),
  gen 56 (scaleTuning). Remove invalid gen 42.
- Implement VibLFO (gen 23/24) with delay and configurable frequency
- Implement ModLFO (gen 21/22) with pitch/filter/volume modulation (gen 5/6/10/11/13)
- SF2 default modulators: velocity→attenuation (concave), velocity→filterFc,
  CC7/CC11→attenuation (concave), CC1→VibLFO
- MIDI: Poly Key Pressure (0xA0), System Common (F1-F3/F6), System Real-Time (F8-FF),
  Channel Pressure storage, readVarLen bounds check, dead SysEx code removal
- SF2GenOper enum corrected to match spec
- presetBagNdx/instBagNdx read as uint32

+++ cpp/src/soundfont.cpp
@@ -132,7 +132,7 @@ bool SoundFont::parsePDTA(const uint8_t* data, size_t size) {
                 m_presets[i].bank = readUint16(d + 22);
                 // SF2 spec: dwPresetBagNdx is at offset 24 (4 bytes)
                 // But some SF2 files use 38-byte records where bagNdx is at offset 24
-                m_presets[i].presetBagNdx = readUint16(d + 24);
+                m_presets[i].presetBagNdx = readUint32(d + 24);
             }
         } else if (memcmp(chunkId, "pbag", 4) == 0) {
             size_t count = chunkSize / 4;
@@ -157,7 +157,7 @@ bool SoundFont::parsePDTA(const uint8_t* data, size_t size) {
                 const uint8_t* d = data + pos + i * 22;
                 m_instruments[i].name = std::string((const char*)d, 20);
                 while (!m_instruments[i].name.empty() && m_instruments[i].name.back() == '\0') m_instruments[i].name.pop_back();
-                m_instruments[i].instBagNdx = readUint16(d + 20); // uint16, not uint32
+                m_instruments[i].instBagNdx = readUint32(d + 20);
             }
         } else if (memcmp(chunkId, "ibag", 4) == 0) {
             size_t count = chunkSize / 4;
